Brief 097: widen CP9 hd-band index counters to int64 (fix HSV M=152K segfault)
cp9_GrowHDBands() accumulates hd_needed = Sum_v (jmax[v]-jmin[v]+1), the
total number of (v,j) band-index cells across all CM states. For HSV
(M=456670 states, L=149643), this sum exceeds 2^31, overflowing the
int32 counters:
  - CP9Bands_t.hd_needed / hd_alloced (infernal.h)
  - local cur_size in cp9_GrowHDBands (the hdmin[v]=hdmin_mem+cur_size
    pointer offset)
The overflow produced an undersized hdmin_mem/hdmax_mem allocation and
wild per-state row pointers; ij2d_bands() then wrote out of bounds -->
SIGSEGV (signal 139) inside cp9_FBMatrices2BandsF, after pinbridge band
derivation had completed successfully.

Widen hd_needed, hd_alloced, cur_size, and the cp9_ValidateBands local
to int64_t. Add a size-gated #HDBANDS stderr diagnostic. SizeofCP9Bands
float accumulation is now 64-bit-safe via the widened hd_alloced.

Latent (not fixed, out of cmalign scope): cp9_CloneBands() passes
hd_alloced to esl_vec_ICopy(int n) -- would truncate at >2^31, but that
path is cmsearch/cmscan-only (cm_pipeline.c), not the cmalign p7band
path that crashes here.

+++ src/infernal.h
@@ -809,8 +809,8 @@ typedef struct cp9bands_s {
   int *safe_hdmax;            /* [0..cm_M-1] safe_hdmax[v] = max_d (hdmax[v][j0]) (over all valid j0) */
 
   /* info on size of bands */
-  int hd_needed;              /* Sum_v cp9b->jmax[v] - cp9b->jmin[v] + 1, number of hd arrays needed */
-  int hd_alloced;             /* number of hd arrays currently alloc'ed */
+  int64_t hd_needed;          /* Sum_v cp9b->jmax[v] - cp9b->jmin[v] + 1, number of hd arrays needed (int64: can exceed 2^31 for very large M*L, e.g. HSV M=152K L=150K) */
+  int64_t hd_alloced;         /* number of hd arrays currently alloc'ed */
 
   double   tau;               /* tau used to calculate current bands */
